ioQ3 Synchro to rev2095: r2087: Fix legacy protocol with new packet queueing r2088: * Use specific exit code for xmessage r2089: Fix delta compression breaking due to packet queuing r2090,2094: Bug 4812 - GCC __attribute__ annotations for printf, non-returning functions etc., patch by linux@youmustbejoking.demon.co.uk and Zack Middleton r2091: - Bug 5083 - Cross compiling for 64bit is missing libcurl.a - Fix a compiler warning for cross compile r2092: * Fix various issues with unix Sys_Dialog r2093: Shut up returning functions with noreturn attribute warning r2095: Bug 5084 - [patch] add support for Debian cross compiling, by q3urt.undead@gmail.com
git-svn-id: https://smokinguns.svn.sourceforge.net/svnroot/smokinguns@592 07eb42f0-b032-4355-afd3-42b1f3832223

+++ branches/1.1/code/renderer/tr_glslprogs.c
@@ -370,7 +370,7 @@ void R_GLSLProgs_Init( void ) {
 		} else {
 			glsl_progs[i].program = glsl_progs[i].default_program ;
 			ri.Printf( PRINT_ALL, "Using default %s GLSL program (%i bytes)\n",
-				glsl_progs[i].name, strlen(glsl_progs[i].program) );
+				glsl_progs[i].name, (int)strlen(glsl_progs[i].program) );
 		}
 
 		if (buffer) {
